+++ forgero-fabric/src/main/java/com/sigmundgranaas/forgero/fabric/resources/dynamic/PartToSchematicGenerator.java
@@ -1,7 +1,7 @@
 package com.sigmundgranaas.forgero.fabric.resources.dynamic;
 
 import com.sigmundgranaas.forgero.core.ForgeroStateRegistry;
-import com.sigmundgranaas.forgero.core.state.Composite;
+import com.sigmundgranaas.forgero.core.state.composite.Construct;
 import net.devtech.arrp.api.RuntimeResourcePack;
 import net.devtech.arrp.json.recipe.*;
 import net.minecraft.item.Items;
@@ -9,6 +9,7 @@ import net.minecraft.util.Identifier;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.function.Supplier;
 
 public class PartToSchematicGenerator implements DynamicResourceGenerator {
 
@@ -17,25 +18,26 @@ public class PartToSchematicGenerator implements DynamicResourceGenerator {
         parts().forEach(comp -> createRecipe(comp).ifPresent(recipe -> pack.addRecipe(id(comp), recipe)));
     }
 
-    private List<Composite> parts() {
+    private List<Construct> parts() {
         return ForgeroStateRegistry.STATES.all().stream()
-                .filter(Composite.class::isInstance)
-                .map(Composite.class::cast)
+                .map(Supplier::get)
+                .filter(Construct.class::isInstance)
+                .map(Construct.class::cast)
                 .filter(comp -> comp.ingredients().stream().anyMatch(ingredient -> ingredient.name().contains("schematic")))
                 .toList();
     }
 
-    private Optional<JRecipe> createRecipe(Composite composite) {
-        var schematic = composite.ingredients().stream().filter(ingredient -> ingredient.name().contains("schematic")).findFirst();
+    private Optional<JRecipe> createRecipe(Construct construct) {
+        var schematic = construct.ingredients().stream().filter(ingredient -> ingredient.name().contains("schematic")).findFirst();
         if (schematic.isPresent()) {
-            var ingredients = JIngredients.ingredients().add(JIngredient.ingredient().item(Items.PAPER)).add(JIngredient.ingredient().item(composite.identifier()));
+            var ingredients = JIngredients.ingredients().add(JIngredient.ingredient().item(Items.PAPER)).add(JIngredient.ingredient().item(construct.identifier()));
             return Optional.of(JShapelessRecipe.shapeless(ingredients, JResult.result(schematic.get().identifier())));
         }
         return Optional.empty();
 
     }
 
-    private Identifier id(Composite composite) {
-        return new Identifier(composite.nameSpace(), composite.name() + "-schematic_recipe");
+    private Identifier id(Construct construct) {
+        return new Identifier(construct.nameSpace(), construct.name() + "-schematic_recipe");
     }
 }