+++ minecraft-common/src/main/java/com/sigmundgranaas/forgero/minecraft/common/item/nbt/v2/CompositeParser.java
@@ -2,8 +2,8 @@ package com.sigmundgranaas.forgero.minecraft.common.item.nbt.v2;
 
 import com.sigmundgranaas.forgero.core.ForgeroStateRegistry;
 import com.sigmundgranaas.forgero.core.registry.StateFinder;
-import com.sigmundgranaas.forgero.core.state.Composite;
 import com.sigmundgranaas.forgero.core.state.State;
+import com.sigmundgranaas.forgero.core.state.composite.Construct;
 import com.sigmundgranaas.forgero.core.type.Type;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtElement;
@@ -25,14 +25,14 @@ public class CompositeParser implements CompoundParser<State> {
         if (!compound.contains(NbtConstants.STATE_TYPE_IDENTIFIER)) {
             return Optional.empty();
         }
-        Composite.CompositeBuilder builder = Composite.builder();
+        Construct.ConstructBuilder builder = Construct.builder();
 
         if (compound.contains(NbtConstants.ID_IDENTIFIER)) {
             var id = compound.getString(NbtConstants.ID_IDENTIFIER);
             var stateOpt = supplier.find(id);
 
-            if (stateOpt.isPresent() && stateOpt.get() instanceof Composite composite) {
-                builder = Composite.builder(composite.slots());
+            if (stateOpt.isPresent() && stateOpt.get() instanceof Construct construct) {
+                builder = Construct.builder(construct.slots());
             } else if (ForgeroStateRegistry.CONTAINER_TO_STATE.containsKey(id)) {
                 return supplier.find(ForgeroStateRegistry.CONTAINER_TO_STATE.get(id));
             }